+++ src/Web/Template/Servant/Auth.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE LambdaCase      #-}
-{-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE ViewPatterns    #-}
+{-# LANGUAGE LambdaCase          #-}
+{-# LANGUAGE OverloadedStrings   #-}
+{-# LANGUAGE RecordWildCards     #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
 module Web.Template.Servant.Auth
   ( CbdAuth
@@ -8,28 +9,31 @@ module Web.Template.Servant.Auth
   , UserId (..)
   , OIDCConfig (..)
   , defaultOIDCCfg
+  , Permit
   ) where
 
 -- after https://www.stackage.org/haddock/lts-15.15/servant-server-0.16.2/src/Servant.Server.Experimental.Auth.html
 
 import           Control.Applicative    ((<|>))
-import           Control.Lens           (At (at), ix, (&), (.~), (<&>), (?~), (^?!))
-import           Control.Monad.Except   (runExceptT)
-import           Control.Monad.IO.Class (liftIO)
-import           Data.IORef             (writeIORef)
+import           Control.Lens           (At (at), ix, (&), (.~), (<&>), (?~), (^..), (^?))
+import           Control.Monad.Except   (runExceptT, unless)
+import           Control.Monad.IO.Class (MonadIO, liftIO)
+import           Data.IORef             (readIORef, writeIORef)
 import           Data.Maybe             (catMaybes)
 import           Data.Proxy             (Proxy (..))
-import           Data.Text              (Text)
+import           Data.Text              (Text, intercalate, pack)
 import           Data.Text.Encoding     (decodeUtf8)
 import qualified Data.Vault.Lazy        as V
 import           GHC.Generics           (Generic)
+import           GHC.TypeLits           (KnownSymbol, Symbol, symbolVal)
 
 import           Crypto.JOSE.JWK                (JWKSet)
-import           Crypto.JWT                     (JWTError, decodeCompact,
-                                                 defaultJWTValidationSettings, unregisteredClaims,
-                                                 verifyClaims)
-import           Data.Aeson                     (Value (..))
-import           Data.ByteString                (stripPrefix)
+import           Crypto.JWT                     (ClaimsSet, JWTError, JWTValidationSettings,
+                                                 SignedJWT, audiencePredicate, decodeCompact,
+                                                 defaultJWTValidationSettings, issuerPredicate,
+                                                 string, unregisteredClaims, uri, verifyClaims)
+import           Data.Aeson.Lens                (AsPrimitive (_String), key, values)
+import           Data.ByteString                (ByteString, stripPrefix)
 import qualified Data.ByteString.Lazy           as LB
 import           Data.Cache                     (Cache)
 import qualified Data.Cache                     as Cache
@@ -38,20 +42,20 @@ import           Data.OpenApi.Internal          (ApiKeyLocation (..), ApiKeyPara
                                                  SecurityScheme (..), SecuritySchemeType (..))
 import           Data.OpenApi.Lens              (components, description, security, securitySchemes)
 import           Data.OpenApi.Operation         (allOperations, setResponse)
-import           Data.Time.Clock                (diffUTCTime, getCurrentTime,
-                                                 nominalDiffTimeToSeconds, UTCTime)
+import           Data.Time.Clock                (NominalDiffTime, UTCTime, addUTCTime, diffUTCTime,
+                                                 getCurrentTime, nominalDiffTimeToSeconds)
 import           Network.HTTP.Client            (Manager, httpLbs)
 import           Network.HTTP.Client.TLS        (newTlsManager)
 import           Network.HTTP.Types.Header      (hContentType)
-import           Network.URI                    (URI)
-import           Network.Wai                    (requestHeaders, vault)
-import           OpenID.Connect.Client.Provider (Discovery (Discovery, jwksUri), keysFromDiscovery)
-import qualified OpenID.Connect.Client.Provider as OIDC
+import           Network.URI                    (URI (..))
+import           Network.Wai                    (Request, requestHeaders, vault)
+import           OpenID.Connect.Client.Provider (Discovery, discovery, keysFromDiscovery)
 import           Servant.API                    ((:>))
 import           Servant.OpenApi                (HasOpenApi (..))
 import           Servant.Server                 (HasContextEntry (getContextEntry), HasServer (..),
-                                                 ServerError (..), err401)
-import           Servant.Server.Internal        (addAuthCheck, delayedFailFatal, withRequest)
+                                                 ServerError (..), err401, err403, err500)
+import           Servant.Server.Internal        (DelayedIO, addAuthCheck, delayedFailFatal,
+                                                 withRequest)
 import           System.Clock                   (TimeSpec (..))
 import           Web.Cookie                     (parseCookiesText)
 
@@ -86,7 +90,7 @@ instance HasServer api context => HasServer (CbdAuth :> api) context where
                 <&> parseCookiesText
                 >>= lookup "id"
           case mUserId of
-            Nothing -> delayedFailFatal err
+            Nothing -> unauth401
             Just uid -> do
                 -- Try to store user id in the vault, to be used by logging middleware later.
                 let mUserIdRef = V.lookup userIdVaultKey $ vault req
@@ -94,11 +98,6 @@ instance HasServer api context => HasServer (CbdAuth :> api) context where
                   Nothing  -> return ()
                   Just ref -> liftIO $ writeIORef ref $ Just uid
                 return $ UserId uid
-    where
-      err = err401
-        { errBody = "{\"error\": \"Authorization failed\"}"
-        , errHeaders = [(hContentType, "application/json")]
-        }
 
 instance HasOpenApi api => HasOpenApi (CbdAuth :> api) where
   toOpenApi _ = toOpenApi @api Proxy
@@ -115,35 +114,49 @@ instance HasOpenApi api => HasOpenApi (CbdAuth :> api) where
 -- Usage:
 --
 -- > type API = OIDCAuth :> (....)
--- Tooks token from 'Authorization' header
--- Handlers will get an 'UserId' argument
--- Stores token and claims in vault
+--
+-- Takes token from 'Authorization' header.
+--
+-- Handlers will get an 'UserId' argument.
+--
+-- Stores token and claims in vault.
 data OIDCAuth
 
 -- | Info needed for OIDC authorization & key cache
-data OIDCConfig = OIDCConfig
-  { oidcManager       :: Manager -- ^ https manager
-  , oidcClientId      :: Text -- ^ audience
-  , oidcIssuer        :: URI -- ^ discovery uri
-  , oidcWorkaroundUri :: URI -- ^ temporary solution to openid-connect issue
-  , oidcKeyCache      :: Cache Text JWKSet -- ^ cache - storing validation keys
-  }
+data OIDCConfig
+  = OIDCConfig
+      { oidcManager           :: Manager
+        -- ^ https manager
+      , oidcClientId          :: Text
+        -- ^ audience
+      , oidcIssuer            :: URI
+        -- ^ discovery uri
+      , oidcDiscoCache        :: Cache () Discovery
+        -- ^ cache - storing discovery information
+      , oidcKeyCache          :: Cache () JWKSet
+        -- ^ cache - storing validation keys
+      , oidcDefaultExpiration :: NominalDiffTime
+        -- ^ Default expiration time for discovery document and JWKS
+      }
 
-defaultOIDCCfg :: IO OIDCConfig
+defaultOIDCCfg :: MonadIO m => m OIDCConfig
 defaultOIDCCfg = do
-  cache <- Cache.newCache (Just 0)
+  discoCache <- liftIO $ Cache.newCache $ Just 0
+  keyCache <- liftIO $ Cache.newCache $ Just 0
   mgr <- newTlsManager
   return $ OIDCConfig
     { oidcManager = mgr
-    , oidcKeyCache = cache
-    , oidcWorkaroundUri = error "workaround uri not set"
+    , oidcDiscoCache = discoCache
+    , oidcKeyCache = keyCache
     , oidcIssuer = error "discovery uri not set"
     , oidcClientId = error "client id not set"
+    , oidcDefaultExpiration = 10 * 60 -- 10 minutes
     }
 
 instance ( HasServer api context
          , HasContextEntry context OIDCConfig
          ) => HasServer (OIDCAuth :> api) context where
+
   type ServerT (OIDCAuth :> api) m = UserId -> ServerT api m
 
   hoistServerWithContext _ pc nt s = hoistServerWithContext @api Proxy pc nt . s
@@ -153,49 +166,22 @@ instance ( HasServer api context
       $ addAuthCheck sub
       $ withRequest $ \req -> do
 
-        token <- case getHeader req of
-          Just token -> return token
-          Nothing    -> delayedFailFatal err
+        token <- maybe unauth401 return $ getToken req
 
-        jws <- case getHeader req >>= getToken of
-          Just jws -> return jws
-          Nothing  -> delayedFailFatal err
+        jwt <- getJWT token
 
-        let OIDCConfig {..} = getContextEntry context
+        let cfg = getContextEntry context
 
-        jwkSet <- liftIO (Cache.lookup oidcKeyCache "jwkSet") >>= \case
-          Just jwkSet -> return jwkSet
-          Nothing     -> liftIO
-              ( keysFromDiscovery
-                (https oidcManager)
-                (Discovery {jwksUri = OIDC.URI $ oidcWorkaroundUri})
-              ) >>= \case
-                Left jwtErr -> do
-                  logWarn $ show jwtErr
-                  delayedFailFatal err
-                Right (jwkSet, mbKeysExp) -> liftIO $ do
-                  now <- getCurrentTime
-                  Cache.insert' oidcKeyCache
-                    (diffTime <$> (mbKeysExp <|> pure now) <*> pure now)
-                    "jwkSet"
-                    jwkSet
-                  return jwkSet
+        disco <- getDisco cfg
 
-        claims <- liftIO
-          ( runExceptT $
-            verifyClaims @_ @_ @JWTError
-              (defaultJWTValidationSettings audCheck)
-              jwkSet
-              jws
-          ) >>= \case
-            Left jwtErr  -> do
-              logWarn $ show jwtErr
-              delayedFailFatal err
-            Right claims -> return claims
+        jwkSet <- getJWKSet cfg disco
 
-        let String uid = claims
-              ^?! unregisteredClaims
-              .ix "object_guid"
+        claims <- getClaims cfg jwt jwkSet
+
+        uid <- maybe
+          (die ERROR unauth401 ("No object_guid found" :: Text))
+          return
+          $ claims ^? unregisteredClaims . ix "object_guid" . _String
 
         liftIO $ sequence_ $ catMaybes
           [ userIdVaultKey <?> req <&> flip writeIORef (Just uid)
@@ -207,29 +193,79 @@ instance ( HasServer api context
     where
       https mgr = (`httpLbs` mgr)
 
-      getHeader r = lookup "Authorization" (requestHeaders r) >>= stripPrefix "Bearer "
+      die :: Show err => Level -> DelayedIO b -> err -> DelayedIO b
+      die lvl fin err = liftIO (log' lvl ("web-template" :: Text) $ show err) >> fin
 
-      getToken = either (const Nothing) Just . decodeCompact @_ @JWTError . LB.fromStrict
+      getToken :: Request -> Maybe ByteString
+      getToken r = lookup "Authorization" (requestHeaders r) >>= stripPrefix "Bearer "
 
-      logWarn = liftIO . log' WARNING ("web-template" :: Text)
-
-      diffTime :: UTCTime -> UTCTime -> TimeSpec
-      diffTime from to = let
-          diff = diffUTCTime from to - tTreshold
-        in max
-            TimeSpec {sec = 0, nsec = 0}
-            TimeSpec {sec = floor $ nominalDiffTimeToSeconds diff, nsec = 0}
+      expiration :: UTCTime -> Maybe UTCTime -> NominalDiffTime -> Maybe TimeSpec
+      expiration now ex defaultExp = diffTime
+          <$> (ex <|> pure (addUTCTime defaultExp now))
+              -- If expiration is not set by OIDC provider, cache data for some
+              -- default amount of time, to avoid too many requests.
+          <*> pure now
         where
           tTreshold = 60 -- consider token expired 'tTreshold' seconds earlier
 
-      audCheck = const True
+          diffTime :: UTCTime -> UTCTime -> TimeSpec
+          diffTime from to = let
+              diff = diffUTCTime from to - tTreshold
+            in max
+                TimeSpec {sec = 0, nsec = 0}
+                TimeSpec {sec = floor $ nominalDiffTimeToSeconds diff, nsec = 0}
 
-      what <?> req = V.lookup what $ vault req
+      getJWT :: ByteString -> DelayedIO SignedJWT
+      getJWT = either (die WARNING unauth401) return . decodeToken
+        where
+            decodeToken = decodeCompact @_ @JWTError . LB.fromStrict
 
-      err = err401
-        { errBody = "{\"error\": \"Authorization failed\"}"
-        , errHeaders = [(hContentType, "application/json")]
-        }
+      getDisco :: OIDCConfig -> DelayedIO Discovery
+      getDisco OIDCConfig {..} = liftIO (Cache.lookup oidcDiscoCache ())
+          >>= maybe
+            fetchDisco
+            return
+        where
+          fetchDisco = liftIO (discovery (https oidcManager) (appWellKnown oidcIssuer))
+              >>= either
+                (die ERROR unauth500)
+                (uncurry discoSuccess)
+            where
+              appWellKnown u@URI {..} = u {uriPath = uriPath <> "/.well-known/openid-configuration"}
+
+              discoSuccess disco mbDiscoExp = liftIO $ do
+                now <- getCurrentTime
+                Cache.insert' oidcDiscoCache (expiration now mbDiscoExp oidcDefaultExpiration) () disco
+                return disco
+
+      getJWKSet :: OIDCConfig -> Discovery -> DelayedIO JWKSet
+      getJWKSet OIDCConfig {..} disco = liftIO (Cache.lookup oidcKeyCache ())
+          >>= maybe
+            fetchKeys
+            return
+        where
+          fetchKeys = liftIO (keysFromDiscovery (https oidcManager) disco)
+              >>= either
+                (die ERROR unauth500)
+                (uncurry keysSuccess)
+            where
+              keysSuccess jwkSet mbKeysExp = liftIO $ do
+                  now <- getCurrentTime
+                  Cache.insert' oidcKeyCache (expiration now mbKeysExp oidcDefaultExpiration) () jwkSet
+                  return jwkSet
+
+      getClaims :: OIDCConfig -> SignedJWT -> JWKSet -> DelayedIO ClaimsSet
+      getClaims OIDCConfig {..} jwt jwkSet = liftIO
+          (runExceptT $
+            verifyClaims @_ @_ @JWTError (jwtValidation oidcIssuer oidcClientId) jwkSet jwt
+          ) >>= either
+            (die ERROR unauth401)
+            return
+        where
+          jwtValidation :: URI -> Text -> JWTValidationSettings
+          jwtValidation issuer audience = defaultJWTValidationSettings (const True)
+            & issuerPredicate .~ (\iss -> iss ^? uri == Just issuer)
+            & audiencePredicate .~ (\aud -> aud ^? string == Just audience)
 
 instance HasOpenApi api => HasOpenApi (OIDCAuth :> api) where
   toOpenApi _ = toOpenApi @api Proxy
@@ -240,3 +276,82 @@ instance HasOpenApi api => HasOpenApi (OIDCAuth :> api) where
       idJWT = SecurityScheme
         (SecuritySchemeHttp $ HttpSchemeBearer $ Just "jwt")
         (Just "jwt token")
+
+-- | Adds authenthication via roles
+--
+-- Usage:
+--
+-- > type API = Permit '["user", "owner"] :> (....)
+-- Route access permitted if user has at least 1 role from specified list
+-- Takes user roles from vault (originated from jwt token)
+data Permit (rs :: [Symbol])
+
+instance ( HasServer api context
+         , KnownSymbols roles
+         , HasContextEntry context OIDCConfig
+         ) => HasServer (Permit roles :> api) context where
+
+  type ServerT (Permit roles :> api) m = ServerT api m
+
+  hoistServerWithContext _ pc nt s = hoistServerWithContext @api Proxy pc nt s
+
+  route _ context sub =
+    route @api Proxy context
+      $ addAuthCheck (const <$> sub)
+      $ withRequest $ \req -> do
+        pTokenRef <- maybe unauth401 return $ pTokenVaultKey <?> req
+
+        claims <- liftIO (readIORef pTokenRef) >>= maybe unauth401 return
+
+        let OIDCConfig {..} = getContextEntry context
+
+        let haveRoles = claims
+              ^.. unregisteredClaims
+              . ix "resource_access"
+              . key oidcClientId
+              . key "roles"
+              . values . _String
+
+        unless (any (`elem` rolesNeeded) haveRoles)
+          unauth403
+    where
+      rolesNeeded = symbolsVal (Proxy @roles)
+
+instance ( HasOpenApi api
+         , KnownSymbols roles
+         ) => HasOpenApi (Permit roles :> api) where
+  toOpenApi _ = toOpenApi @api Proxy
+    & setResponse 403 (return $ mempty & description .~ descr)
+    where
+      descr = "Action not permitted. Allowed for: "
+        <> intercalate ", " (symbolsVal (Proxy :: Proxy roles))
+
+class KnownSymbols (rs :: [Symbol]) where
+  symbolsVal :: p rs ->  [Text]
+
+instance KnownSymbols '[] where
+  symbolsVal _ = []
+
+instance (KnownSymbol h, KnownSymbols t) => KnownSymbols (h ': t) where
+  symbolsVal _ = pack (symbolVal (Proxy :: Proxy h)) : symbolsVal (Proxy :: Proxy t)
+
+(<?>) :: V.Key a -> Request -> Maybe a
+what <?> req = V.lookup what $ vault req
+
+unauth401 :: DelayedIO a
+unauth401 = delayedFailFatal $ err401
+  { errBody = "{\"error\": \"Authorization failed\"}"
+  , errHeaders = [(hContentType, "application/json")]
+  }
+
+unauth403 :: DelayedIO a
+unauth403 = delayedFailFatal $ err403
+  { errBody = "{\"error\": \"Action not permitted\"}"
+  , errHeaders = [(hContentType, "application/json")]
+  }
+
+unauth500 :: DelayedIO a
+unauth500 = delayedFailFatal $ err500
+  { errBody = "{\"error\": \"Internal server error\"}"
+  , errHeaders = [(hContentType, "application/json")]
+  }